Create `Testnet.Components.DReps` module with DRep related functionality
